+++ extensions/sns/src/validate_config.rs
@@ -3,17 +3,19 @@ use fn_error_context::context;
 use std::ffi::OsString;
 use std::path::Path;
 
-use dfx_extensions_utils::call_bundled;
+use dfx_extensions_utils::call_extension_bundled_binary;
 
 /// Checks whether an SNS configuration file is valid.
 #[context("Failed to validate SNS config at {}.", path.display())]
-pub fn validate_config(dfx_cache_path: &Path, path: &Path) -> anyhow::Result<String> {
+pub fn validate_config(path: &Path, dfx_cache_path: &Path) -> anyhow::Result<String> {
     let args = vec![
         OsString::from("init-config-file"),
         OsString::from("--init-config-file-path"),
         OsString::from(path),
         OsString::from("validate"),
     ];
-    call_bundled(dfx_cache_path, "sns", args)
+    // current binary
+
+    call_extension_bundled_binary("sns-cli", &args, dfx_cache_path)
         .map(|_| format!("SNS config file is valid: {}", path.display()))
 }